requirements.txt
tensorflow~=2.18.0
numpy~=2.0.2
matplotlib~=3.10.0
seaborn~=0.13.2
scikit-learn~=1.6.0
pillow~=11.0.0
pandas~=2.2.3
joblib~=1.4.2
pydot~=3.0.3
graphviz~=0.20.3